bug fixed: migration integer column chat_id changed to big integer column

+++ database/migrations/2022_10_09_231130_create_telegram_users_table.php
@@ -14,7 +14,7 @@ class CreateTelegramUsersTable extends Migration
     public function up()
     {
         Schema::create('telegram_users', function (Blueprint $table) {
-            $table->integer('chat_id')->primary();
+            $table->bigInteger('chat_id')->primary();
             $table->foreignId('user_id')->nullable()->references('id')->on('users')->cascadeOnUpdate()->nullOnDelete();
             $table->string('fullname');
             $table->string('current_state')->nullable();
